Refactored udp client and server to use libhangman. Libhangman refactored
