+++ libhangman/buildLinux.sh
@@ -1,5 +1,5 @@
 # Build script for Linux systems
 
-gcc -c -Wall -Werror -fpic game.c
+gcc -c -Wall -Werror -fpic hangman.c
 gcc -shared -o libs/libhangman.so game.o -L../libsocket/libs -lsocket
 sudo cp libs/libhangman.so /usr/lib
