--- udp/udp_client/includes/definitions.h
@@ -1,13 +1,0 @@
-/*
- * Definitions.h
- *
- *  Created on: 4 Oct 2015
- *      Author: david
- */
-
-#ifndef INCLUDES_DEFINITIONS_H_
-#define INCLUDES_DEFINITIONS_H_
-
-#define HANGMAN_TCP_PORT 1071
-
-#endif /* INCLUDES_DEFINITIONS_H_ */
