+++ udp/udp_client/src/makefile
@@ -1,5 +1,5 @@
 GCC=gcc
-LIBS=-lsocket
+LIBS=-lsocket -lhangman
 LIB_PATH=-L"/usr/lib"
 SOURCE_FILES=udp_client.c
